feat: rename behaviour to behavior BREAKING CHANGE: fc.behaviour.* -> fc.behavior.*
