+++ src/behavior/pointer.js
